Sound Update
Got Sound to work by playing music based on the game ball and setting ear receive to the camera position. When the ball moves away from the player the sound decays over time, etc. I'm going to try my Animation on the RVR labs to see if the issue was my computer. I legit followed all the steps but it doesn't seem to work on my computer.